+++ pubspec.yaml
@@ -16,6 +16,8 @@ dependencies:
   shared_preferences: ^0.4.2
   cloud_firestore: ^0.7.4
   firebase_core: ^0.2.5
+  connectivity:
+  fluro: ^1.3.4
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
